Refactor edge connection dialogue
